Actualizaciones en la carpeta evaluaciones

+++ DESARROLLO/SAV/BACKEND/sav-backend/src/main/java/pe/edu/savbackend/domain/tarea/ExamenDto.java
@@ -18,7 +18,7 @@ public class ExamenDto {
 	private String cantidadPreguntas;
 	private String fechaInico;
 	private String horaInicio;
-	private Integer duracion; // en minutos
+	private Double duracion; // en minutos
 
 	private LocalDateTime fechaSolucion;
 	private String fechaRealizada;
